Fix Rails loading
