--- lib/literal/rails.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "railtie"
-require_relative "rails/patches/active_record"
-
-module Literal::Rails
-end
