+++ lib/literal/rails/active_record_relation_patch.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Literal::Rails
+	module ActiveRecordRelationPatch
+		def Relation(model)
+			RelationType.new(model)
+		end
+	end
+end
